requirements.txt
amqp==5.2.0
anyio==4.2.0
APScheduler==3.10.4
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.4.1
autopep8==2.1.0
azure-core==1.26.3
azure-storage-blob==12.14.1
Babel==2.14.0
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
Cerberus==1.3.5
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.2.0
cron-descriptor==1.4.3
cryptography==41.0.5
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==4.0.1
Django==4.2.4
django-allauth==0.54.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-cors-headers-multi==1.2.0
django-extensions==3.2.3
django-filter==23.3
django-redis==5.4.0
django-silk==5.1.0
django-storages==1.13.2
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docx-mailmerge2==0.8.0
docxcompose==1.4.0
docxtpl==0.16.7
drf-yasg==1.21.7
executing==2.0.1
fonttools==4.49.0
geographiclib==2.0
geopy==2.4.1
gprof2dot==2022.7.29
gunicorn==21.2.0
h11==0.14.0
html2docx==1.6.0
html5lib==1.1
htmldocx==0.0.6
httpcore==1.0.3
httpx==0.26.0
idna==3.4
inflection==0.5.1
ipython==8.21.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.3
jsonfield==2.1.1
kiwisolver==1.4.5
kolo==2.18.0
kombu==5.3.5
logtail-python==0.2.10
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
msgpack==1.0.7
msrest==0.7.1
numpy==1.26.4
oauthlib==3.2.2
optional-django==0.3.0
oscrypto==1.3.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
Pillow==10.0.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycountry==23.12.11
pycparser==2.21
Pygments==2.17.2
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyparsing==3.1.1
pypdf==4.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
react==4.3.0
redis==5.0.1
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
svglib==1.5.1
tinycss2==1.2.1
toolz==0.12.1
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ulid-py==1.1.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
whitenoise==6.5.0
xhtml2pdf==0.2.15
